test: extract sdk-specific tests from InitialActivityTest
It could trigger an `UnsatisfiedLinkError` on `openBackend` for all
other tests in the fork, if the test was ordered as the first in the
sequence.

The `@Config` annotation meant that tests were run in a different
Robolectric sandbox.

`RobolectricTest` loads the backend library into the current sandbox
(a separate classloader).
An on-disk backend library file can be owned by a single classloader:
if this test ran first, it loaded the backend library into the
Robolectric-specific sandbox, and other tests in the forked Gradle
process failed with `UnsatisfiedLinkError`.

Tests were moved to fix this: there was no hard requirement on the
backend or `RobolectricTest`.

Part of 14796
